+++ data/handleDB.php
@@ -66,19 +66,25 @@ class DbBackupCommand extends Command
     {
         $this->setName('db:backup')
             ->setDescription('Backup the current SQLite compatinfo database')
+            ->addArgument(
+                'source',
+                InputArgument::REQUIRED,
+                'database source file to copy'
+            )
         ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $source  = $this->getApplication()->getDbFilename();
+        $source  = $this->getApplication()->getAppTempDir() . '/' . $input->getArgument('source');
         $tempDir = $this->getApplication()->getAppTempDir() . '/backups';
 
         if (!file_exists($tempDir)) {
-            mkdir($tempDir, 0755, true);
+            mkdir($tempDir);
         }
+
         $sha1 = sha1_file($source);
-        $dest = $tempDir . '/' . basename($source) . ".$sha1";
+        $dest = tempnam($tempDir, 'db');
 
         copy($source, $dest);
 
@@ -89,6 +95,41 @@ class DbBackupCommand extends Command
                 $source,
                 $sha1,
                 $dest
+            )
+        );
+    }
+}
+
+/**
+ * Clean copies of the database
+ */
+class DbCleanCommand extends Command
+{
+    protected function configure()
+    {
+        $this->setName('db:clean')
+            ->setDescription('Clean all copies of SQLite compatinfo database')
+        ;
+    }
+
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $source  = $this->getApplication()->getAppTempDir();
+        $count   = 0;
+
+        if (file_exists($source)) {
+            foreach (glob("$source/db*.tmp") as $file) {
+                unlink($file);
+                $count++;
+            }
+        }
+
+        $output->writeln(
+            sprintf(
+                '<info>%d</info> database copie(s)' .
+                ' were removed from <comment>%s</comment>',
+                $count,
+                $source
             )
         );
     }
@@ -144,23 +185,8 @@ class DbInitCommand extends Command
             $extensions = array($extension);
         }
 
-        // do a DB backup first
-        $command = $this->getApplication()->find('db:backup');
-        $arguments = array(
-            'command' => 'db:backup',
-        );
-        $input = new ArrayInput($arguments);
-        $returnCode = $command->run($input, $output);
+        $pdo = Environment::initRefDb(true);
 
-        if ($returnCode !== 0) {
-            $output->writeln('<error>DB backup not performed</error>');
-            return;
-        }
-
-        // then delete current DB before to init a new copy again
-        unlink($this->getApplication()->getDbFilename());
-
-        $pdo = new \PDO('sqlite:' . $this->getApplication()->getDbFilename());
         $ref = new ReferenceCollection($pdo);
 
         $max = count($extensions);
@@ -241,7 +267,7 @@ class DbInitCommand extends Command
                 'build_version' => $this->getApplication()->getVersion(),
             )
         );
-        $progress->setMessage('Database is built');
+        $progress->setMessage('Database is built to file ' . Environment::getDbFilename());
         $progress->display();
         $progress->finish();
         $output->writeln('');
@@ -873,6 +899,7 @@ class DbHandleApplication extends Application
         $defaultCommands = parent::getDefaultCommands();
 
         $defaultCommands[] = new DbBackupCommand();
+        $defaultCommands[] = new DbCleanCommand();
         $defaultCommands[] = new DbInitCommand();
         $defaultCommands[] = new DbReleaseCommand();
 
@@ -881,10 +908,7 @@ class DbHandleApplication extends Application
 
     public function getDbFilename()
     {
-        $database = 'compatinfo.sqlite';
-        $source   = __DIR__ . '/' . $database;
-
-        return $source;
+        // @deprecated : do not use anymore
     }
 
     public function getAppTempDir()
@@ -900,6 +924,7 @@ class DbHandleApplication extends Application
     public function getLongVersion()
     {
         $v = Environment::versionRefDb();
+        unlink(Environment::getDbFilename());
 
         return sprintf(
             '<info>%s</info> version <comment>%s</comment> DB built <comment>%s</comment>',
@@ -910,5 +935,5 @@ class DbHandleApplication extends Application
     }
 }
 
-$application = new DbHandleApplication('Database handler for CompatInfo', '1.18.0');
+$application = new DbHandleApplication('Database handler for CompatInfo', '1.19.0');
 $application->run();